Add a (metric) ton of instrumentation
This prints all the variables used, files changed,
commits included in the action as well as the responses
from the helix runtime.

Error isolation is used, so if one request to the helix runtime
fails, this doesn't affect the requests for the rest of the files.

Errors are buffered so multiple errors can be reported to
gh.

Debug output is printed to stderr to ensure it's included.

Various sanity checks are performed.

Some files (those starting with .github/) are ignored, because issuing
HLXPURGE requests for those urls yields a 403 FORBIDDEN status code and
HTML payload from the backend. Such errors should now be caught more
broadly by the error isolation facilities also introduced with this
commit.
This could have led to HLXPURGE requests being omitted if the one of
those files whose HLXPURGE yield 403s was changed along with other files
in the same PR and > 20 files where changed.
There is currently no evidence to support that this ever happened in the
wild.